+++ simulator/core/perf_sim.h
@@ -66,14 +66,14 @@ private:
 
     std::unique_ptr<WritePort<BPInterface>> wp_memory_2_bp = nullptr;
 
-    std::unique_ptr<WritePort<uint64>> wp_execute_2_execute_bypass = nullptr;
-    std::unique_ptr<WritePort<uint64>> wp_memory_2_execute_bypass = nullptr;
+    std::unique_ptr<WritePort<uint128>> wp_execute_2_execute_bypass = nullptr;
+    std::unique_ptr<WritePort<uint128>> wp_memory_2_execute_bypass = nullptr;
 
     static constexpr const uint8 SRC_REGISTERS_NUM = 2;
     static constexpr const uint8 BYPASSING_STAGES_NUM = DataBypass::RegisterStage::get_bypassing_stages_number();
     static constexpr const uint8 BYPASSING_UNIT_FLUSH_NOTIFIERS_NUM = 2;
 
-    std::array<std::array<std::unique_ptr<ReadPort<uint64>>, BYPASSING_STAGES_NUM>, SRC_REGISTERS_NUM> 
+    std::array<std::array<std::unique_ptr<ReadPort<uint128>>, BYPASSING_STAGES_NUM>, SRC_REGISTERS_NUM> 
         rps_stages_2_execute_sources_bypass;
 
     std::array<std::unique_ptr<WritePort<DataBypass::BypassCommand>>, SRC_REGISTERS_NUM> wps_decode_2_execute_command;
